requirements.txt
streamlit
Pillow
openai
requests
beautifulsoup4

